Setting the variable imports back to caps and 'moving' lowercase files to uppercase files in components/ so that we are following proper react conventions
