+++ src/components/WishlistPublic.js
