+++ src/components/__tests__/App.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import App from '../app.js';
+import App from '../App.js';
 import renderer from 'react-test-renderer';
 import { MemoryRouter } from 'react-router-dom'; {/* For testing, otherwise snapshot will fail */}
