+++ src/components/__tests__/Dashboard.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Dashboard from '../dashboard.js';
+import Dashboard from '../Dashboard.js';
 import renderer from 'react-test-renderer';
 
 describe('Dashboard', () => {
