+++ src/components/__tests__/Pairing.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Pairing from '../pairing.js';
+import Pairing from '../Pairing.js';
 import renderer from 'react-test-renderer';
 
 describe('Pairing', () => {
